+++ BPS/src/main/java/com/sales/BPS/mproduct/entity/ApprovalCode.java
@@ -1,0 +1,25 @@
+package com.sales.BPS.mproduct.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.List;
+
+@Entity
+@Data
+@Table(name = "APPROVALCODE")
+public class ApprovalCode {
+
+    @Id
+    @Column(name = "app_code")
+    private String appCode;
+
+    @Column(name = "app_name")
+    private String appName;
+
+    @OneToOne(mappedBy = "approvalCode")
+    private Voucher voucher;
+
+}