+++ BPS/src/main/java/com/sales/BPS/msales/entity/Ppc.java
@@ -1,0 +1,32 @@
+package com.sales.BPS.msales.entity;
+
+import com.sales.BPS.mproduct.entity.Product;
+import jakarta.persistence.*;
+import lombok.Data;
+
+
+@Entity
+@Data
+@IdClass(PpcPK.class)
+@Table(name = "PPC")
+public class Ppc {
+
+    @Id
+    @Column(name = "client_code")
+    private String clientCode;
+
+    @Id
+    @Column(name = "pro_code")
+    private Integer proCode;
+
+    @ManyToOne
+    @JoinColumn(name = "client_code", insertable = false, updatable = false)
+    private Client client;
+
+    @ManyToOne
+    @JoinColumn(name = "pro_code", insertable = false, updatable = false)
+    private Product product;
+
+    @Column(name = "ppc_sale")
+    private Integer ppcSale;
+}