+++ BPS/src/main/java/com/sales/BPS/msales/entity/PpcPK.java
@@ -1,0 +1,10 @@
+package com.sales.BPS.msales.entity;
+
+import lombok.Data;
+
+import java.io.Serializable;
+@Data
+public class PpcPK implements Serializable {
+    private String clientCode;
+    private Integer proCode;
+}