+++ BPS/src/main/java/com/sales/BPS/msystem/entity/AuthorityCode.java
@@ -1,0 +1,22 @@
+package com.sales.BPS.msystem.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+@Entity
+@Data
+@Table(name = "AUTHORITYCODE")
+public class AuthorityCode {
+
+    @Id
+    @Column(name = "auth_code")
+    private String authCode;
+
+    @Column(name = "auth_name")
+    private String authName;
+
+    @OneToOne(mappedBy = "authorityCode")
+    private Authority authority;
+
+}
+