+++ BPS/src/main/java/com/sales/BPS/msystem/entity/AuthorityPK.java
@@ -1,0 +1,13 @@
+package com.sales.BPS.msystem.entity;
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+@Data
+public class AuthorityPK implements Serializable {
+
+    private Integer empCode;
+
+    private String authCode;
+}