+++ BPS/src/main/java/com/sales/BPS/msystem/entity/TempLog.java
@@ -1,0 +1,33 @@
+package com.sales.BPS.msystem.entity;
+
+import jakarta.persistence.*;
+import lombok.Data;
+
+import java.time.LocalDateTime;
+
+import static jakarta.persistence.FetchType.LAZY;
+
+@Entity
+@Data
+@IdClass(TempLogPK.class)
+@Table(name = "TEMPLOG")
+public class TempLog {
+
+    @Id
+    @Column(name = "temp_code")
+    private String tempCode;
+
+    @Id
+    @Column(name = "storage_code")
+    private String storageCode; // 직접 선언
+
+    @ManyToOne(fetch = LAZY)
+    @JoinColumn(name = "storage_code", insertable = false, updatable = false)
+    private Storage storage;
+
+    @Column(name = "temp_temp")
+    private Integer tempTemp;
+
+    @Column(name = "temp_date")
+    private LocalDateTime tempDate;
+}