+++ BPS/src/main/java/com/sales/BPS/msystem/repository/AuthorityCodeRepository.java
@@ -1,0 +1,7 @@
+package com.sales.BPS.msystem.repository;
+
+import com.sales.BPS.msystem.entity.AuthorityCode;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthorityCodeRepository extends JpaRepository<AuthorityCode, String> {
+}