+++ BPS/src/main/java/com/sales/BPS/msystem/repository/AuthorityRepository.java
@@ -1,0 +1,7 @@
+package com.sales.BPS.msystem.repository;
+
+import com.sales.BPS.msystem.entity.Authority;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthorityRepository extends JpaRepository<Authority, String> {
+}