+++ BPS/src/main/java/com/sales/BPS/msystem/repository/EmployeeRepository.java
@@ -1,0 +1,7 @@
+package com.sales.BPS.msystem.repository;
+
+import com.sales.BPS.msystem.entity.Employee;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EmployeeRepository extends JpaRepository<Employee, String> {
+}