+++ changelog.md
@@ -5,5 +5,45 @@
 2024-03-19 AM 10:18
 - [fix] yml 파일
 
+2024-03-20 AM 10:10
+- [add] dev 브랜치 추가
+
+2024-03-20 AM 11:40
+- [add] 디렉토리 추가 및 test파일 생성
+
+2024-03-20 AM 11:57
+- [remove] test파일 삭제
+
+2024-03-20 PM 12:05
+- [add] test 파일 재생성
+
+0.1.0/4.1.1/SH_SM
+2024-03-20 PM 12:39
+- [feat] 0.1.0/4.1.1/SH_SM spring-security 기본 구성 생성
+
+0.1.0/4.5.1/MS
+2024-03-20 PM :14
+- [Feat] 0.1.0/4.5.1/MS PRODUCT, STOCK  ENTITY생성
+
+2024-03-20 PM 12:14
+- [Feat] 0.1.0/4.10.1/WY 매출거래처 등록 브랜치 생성 
+
+2024-03-20 PM 03:17
+- [Feat] 0.1.1/4.1.1/SH_SM spring-security 브랜치 생성
+
+0.1.0/4.3.2/SM
+- [add] 알림 관련 Eneity 생성
+
+0.1.0/4.7.1/MS
+2024-03-20 PM 15:09
+- [Feat] 0.1.0/4.7.1/MS VOUCHER ENTITY 생성
+
+2024-03-20 PM 03:07
+- [Feat] 0.1.0/4.11/WY 콜드체인관련 ENTITY 생성
+
+2024-03-20 PM 04:23
+- [fix] Client, Ppc ENTITY 수정
+
 2024-03-21
 0.1.0/release
+- [Feat] deploy aws ec2